changing explicit mr representation
